Downgrade Ansible to 2.16.x so we can manage Rocky8 hosts
Starting at version 2.17, Ansible can no longer use Python 3.6 so it
can't manage Rocky8 hosts unless we install a full Python 3.9
environment, which seems like a lot of work.

For now I'll use this EE for everything, but if we find that it can't
work with our playbook then I'll fork it and have one EE for Rocky8
and one for everything else.